requirements.txt
Babel==1.3
cliff==1.13.0
cmd2==0.6.8
configobj==5.0.6
configparser==3.3.0.post2
debtcollector==0.5.0
decorator==4.0.2
ecdsa==0.13
Fabric==1.10.2
futures==3.0.3
Gluttony==0.8
iso8601==0.1.10
logging==0.4.9.6
monotonic==0.2
msgpack-python==0.4.6
multiprocessing==2.6.2.1
netaddr==0.7.15
netifaces==0.10.4
networkx==1.10
oslo.config==1.15.0
oslo.i18n==2.1.0
oslo.serialization==1.7.0
oslo.utils==1.9.0
paramiko==1.15.2
pbr==1.3.0
pexpect==3.3
prettytable==0.7.2
pudb==2015.3
pycrypto==2.6.1
Pygments==2.0.2
pyparsing==2.0.3
python-group-based-policy-client==0.9.2
python-heatclient==0.6.0
python-keystoneclient==1.6.0
python-neutronclient==2.6.0
python-swiftclient==2.4.0
pytz==2015.4
PyYAML==3.11
requests==2.7.0
simplejson==3.7.3
six==1.9.0
stevedore==1.6.0
urwid==1.3.0
wrapt==1.10.5
